complete update blog controller and fix bug multer

+++ app/utils/functions.js
@@ -67,8 +67,10 @@ function verifyRefreshToken(token) {
 }
 
 function deleteFileInPublic(fileAddress) {
-  const pathFile = path.join(__dirname, "..", "..", "public", fileAddress);
-  fs.unlinkSync(pathFile);
+  if (fileAddress) {
+    const pathFile = path.join(__dirname, "..", "..", "public", fileAddress);
+    if (fs.existsSync(pathFile)) fs.unlinkSync(pathFile);
+  }
 }
 
 module.exports = {
